Added Poetry.lock file to include all dependencies
The Poetry.lock file is now added to specify all the dependencies needed for the application. This file lists in detail all the packages required along with their versions and descriptions. It helps maintain consistency across different environments where the application is being run. If any new dependency is added to the project or any dependency version is upgraded, this file needs to be updated. The Poetry.lock file can be updated using the 'poetry lock' command in the terminal.